refactor: rename to mongo.py and add logging module
